requirements.txt
gradio
pyyaml~=6.0
pandas
requests